MultiThread implementation of preprocessor
This is a multithread implementation of the song preprocessor. There is
an issue with python's GIL that makes it not a true multithreaded
application. I created this in a separate branch because I'm not
entirely sure how appropriate it would be to push to master. Potential
also for creating a multiPROCESS app.